Correct error in sending commands to device
I thought that commands sent to the device were prefixed with a null
byte. After checking the actual HID reports, it appears that this is not
the case.

This means that the data can contain one more byte (up to 52 bytes now).

Tests have been updated and the length of the HID report was checked
using USB monitoring (`ussmon`).